[Silver III] Title: 주유소, Time: 20 ms, Memory: 3460 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -9,7 +9,7 @@ int main() {
     long long sum = 0;
     cin >> cnt;
 
-    int dist[100000], price[100000];
+    long long dist[100000], price[100000];
 
     for (int i = 0; i < cnt - 1; ++i) {
         cin >> dist[i];
